+++ examples/two-boxes.rs
@@ -4,8 +4,8 @@
 use std::io::stdout;
 
 use gerber_types::{
-    Aperture, ApertureDefinition, Circle, Command, CoordinateFormat, Coordinates, DCode,
-    ExtendedCode, FileAttribute, FunctionCode, GCode, GenerationSoftware, GerberCode,
+    Aperture, ApertureDefinition, Circle, Command, CommentContent, CoordinateFormat, Coordinates,
+    DCode, ExtendedCode, FileAttribute, FunctionCode, GCode, GenerationSoftware, GerberCode,
     InterpolationMode, MCode, Operation, Part, Polarity, Unit,
 };
 
@@ -14,7 +14,10 @@ const VERSION: &'static str = env!("CARGO_PKG_VERSION");
 fn main() {
     let cf = CoordinateFormat::new(2, 6);
     let commands: Vec<Command> = vec![
-        FunctionCode::GCode(GCode::Comment("Ucamco ex. 1: Two square boxes".to_string())).into(),
+        FunctionCode::GCode(GCode::Comment(CommentContent::String(
+            "Ucamco ex. 1: Two square boxes".to_string(),
+        )))
+        .into(),
         ExtendedCode::Unit(Unit::Millimeters).into(),
         ExtendedCode::CoordinateFormat(cf).into(),
         ExtendedCode::FileAttribute(FileAttribute::GenerationSoftware(GenerationSoftware::new(